// projects/cd-common/src/lib/components/data-picker/data-tree/data-tree.interfaces.ts
/*
 * Copyright 2021 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { ComponentSize } from 'cd-interfaces';

export type ProcessedData = [cells: IObjectNode[], expanded: string[]];

export interface IObjectIcon {
  name?: string;
  size?: ComponentSize;
}

export interface IObjectNode {
  id: string;
  type?: string;
  title: string;
  level: number;
  selectable: boolean;
  pos: number[];
  children?: boolean;
  parent?: string;
  rootId?: string;
  icon?: IObjectIcon;
  value?: number | boolean | undefined | string;
  inputs?: boolean;
}